#323 [feat]: 나의 답변 모아보기 화면 내 하트 클릭 기능 구현

+++ app/src/main/java/com/byeboo/app/presentation/quest/review/common/personal/MyAnswerScreen.kt
@@ -73,7 +73,7 @@ fun MyAnswerScreen(
     onBackClick: () -> Unit,
     onMyAnswerContentClick: (Long) -> Unit,
     paddingValues: PaddingValues,
-    onHeartClick: () -> Unit,
+    onHeartClick: (Long) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     val listState = rememberLazyListState()
